Changed Lang file structure. Abstraction of MainActivity.kt. Added proguard-rules.pro for Lang.model
