+++ app/src/main/java/vfunny/shortvideovfunnyapp/Lang/model/language.kt
@@ -1,4 +1,4 @@
-package vfunny.shortvideovfunnyapp.Data.model
+package vfunny.shortvideovfunnyapp.Post.model
 
 import android.content.Context
 import android.util.Log
