+++ app/src/main/java/vfunny/shortvideovfunnyapp/Live/ui/ListActivity.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import coil.imageLoader
 import com.google.firebase.database.*
-import vfunny.shortvideovfunnyapp.Data.model.Post
+import vfunny.shortvideovfunnyapp.Post.model.Post
 import vfunny.shortvideovfunnyapp.databinding.ListActivityBinding
