+++ app/src/main/java/vfunny/shortvideovfunnyapp/Login/model/User.kt
@@ -2,8 +2,8 @@ package vfunny.shortvideovfunnyapp.Login.model
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.*
-import vfunny.shortvideovfunnyapp.Data.model.Const
-import vfunny.shortvideovfunnyapp.Data.model.Language
+import vfunny.shortvideovfunnyapp.Post.model.Const
+import vfunny.shortvideovfunnyapp.Post.model.Language
 
 /**
  * Created on 02/05/2019.
