+++ app/src/main/java/vfunny/shortvideovfunnyapp/Post/model/Const.kt
@@ -1,4 +1,4 @@
-package vfunny.shortvideovfunnyapp.Data.model
+package vfunny.shortvideovfunnyapp.Post.model
 
 /**
  * Created on 06/07/2019.
